+++ src/app/tests/landing-page.component.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AppComponent } from './app.component';
+import { AppComponent } from '../components/landing-page/landing-page.component';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
